ID to all btns, move config files to common dir, misc test fixes
