+++ frontend/components/header/HeaderWebsite.vue
@@ -9,7 +9,7 @@
       'invisible opacity-0': headerOpacity == 0,
     }"
   >
-    <!-- Note: Mobile header including SidebarRight. -->
+    <!-- MARK: Mobile Header -->
     <div id="mobile-header" class="flex px-4 py-3 md:hidden">
       <div class="z-0 mx-auto">
         <div
@@ -36,7 +36,7 @@
         </div>
       </SidebarRight>
     </div>
-    <!-- Note: Desktop header. -->
+    <!-- MARK: Desktop Header -->
     <div id="desktop-header" class="mx-auto hidden py-3 md:block">
       <div class="responsive-px-5 flex items-center justify-between">
         <div class="flex items-center md:space-x-4 lg:space-x-6 xl:space-x-8">
