+++ frontend/nuxt.config/head.ts
