+++ frontend/nuxt.config/locales.ts
