+++ frontend/nuxt.config/modules.ts
