+++ frontend/pages/events/create.vue
@@ -267,6 +267,7 @@
             type="button"
             class="absolute left-0 mr-2"
             :class="{ 'placeholder-class': hasPreviousPage }"
+            :cta="false"
             label="pages.events.create.button-left"
             leftIcon="←"
             fontSize="lg"
@@ -287,6 +288,7 @@
             @click="nextPage"
             type="button"
             class="absolute right-0 ml-2"
+            :cta="false"
             label="pages.events.create.button-right"
             rightIcon="→"
             fontSize="lg"
