+++ frontend/types/btn-props.d.ts
@@ -1,5 +1,6 @@
 export interface BtnBase {
-  cta?: boolean;
+  id?: string;
+  cta: boolean;
   label?: string;
   hideLabelOnMobile?: boolean;
   fontSize: "xs" | "sm" | "base" | "lg" | "xl" | "2xl" | "3xl";
@@ -9,7 +10,6 @@ export interface BtnBase {
 }
 
 withDefaults(defineProps<BtnBase>(), {
-  cta: false,
   fontSize: "base",
   iconSize: "1em",
 });
@@ -31,5 +31,4 @@ export interface BtnRoute extends BtnBase {
   linkTo: string;
   leftIcon?: string;
   rightIcon?: string;
-  id?: string;
 }
